- Added variantId in OrderItemRequest - Added custom query to get variant according to productId and variantId - Endpoint to check product variant availability - Endpoint to decrease stock after successful order creation - Added productVariantId and quantity in OrderItem - Updated thoughtProcess.md

+++ product-service/src/main/java/com/encora/purab/product_service/entity/Product.java
@@ -27,7 +27,7 @@ public class Product {
     @Transient
     private int stock;
 
-    public double getStock(){
+    public int getStock(){
         int stock = 0;
         for(ProductVariant productVariant : productVariantList){
             stock += productVariant.getStock();
